requirements.txt
# LangGraph core
langgraph>=0.2.0
langchain-core>=0.3.0

# LLM providers
langchain-openai>=0.2.0

# Config
pyyaml>=6.0
